# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GovCon AI Scanner API
  description: |
    # GovCon AI Scanner — REST API Reference

    The GovCon AI Scanner helps government contractors discover, analyze, and
    track opportunities from SAM.gov.  Use this API to integrate opportunity
    data into your ERP or build custom workflows.

    ## Authentication

    All protected endpoints require a **Bearer** access token obtained via
    `POST /api/auth/login`.  Tokens expire after **15 minutes**; use
    `POST /api/auth/refresh` with your refresh token to obtain a new pair.

    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting

    | Endpoint group        | Limit                 |
    |-----------------------|-----------------------|
    | `/api/auth/*`         | 30 requests / 15 min  |
    | `/api/opportunities/*`| 200 requests / 15 min |
    | `/api/email*`         | 200 requests / 15 min |
    | `/api/admin/*`        | 60 requests / 15 min  |

    When exceeded the API returns **429 Too Many Requests** with:
    ```json
    { "success": false, "error": "Too many requests. Please try again later." }
    ```

    ## Error Codes

    | HTTP Code | Meaning |
    |-----------|---------|
    | 400       | Bad Request — missing or invalid fields |
    | 401       | Unauthorized — missing, expired, or invalid token |
    | 403       | Forbidden — valid token but insufficient permissions |
    | 404       | Not Found |
...
    | 429       | Too Many Requests |
...

...

...

...

...
  contact:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "64a7b2c3d4e5f6a7b8c9d0e1"
        email:
          type: string
          format: email
          example: "jane@example.com"
        name:
          type: string
          example: "Jane Smith"
        company:
          type: string
          example: "Acme Federal LLC"
        naicsCodes:
          type: array
          items:
            type: string
          example: ["541511", "541330"]
        role:
          type: string
          enum: [user, admin]
          example: user
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UserIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/admin/users:
    get:
      tags: [Admin – Users]
      summary: List all users
      description: Paginated, filterable list of all user accounts. **Admin only.**
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          description: Search by email, name, or company
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /api/admin/users/{id}:
    get:
      tags: [Admin – Users]
      summary: Get user details and usage stats
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: User detail
        "404":
          description: User not found

    patch:
      tags: [Admin – Users]
      summary: Update user (admin override)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                company:
                  type: string
                naicsCodes:
                  type: array
                  items:
                    type: string
                role:
                  type: string
                  enum: [user, admin]
                isActive:
                  type: boolean
      responses:
        "200":
          description: User updated

    delete:
      tags: [Admin – Users]
      summary: Deactivate user (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: User deactivated
  /api/admin/users/{id}/activity:
    get:
      tags: [Admin – Users]
      summary: User activity log
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Activity log

  # ==========================================================================
  # ADMIN – ANALYTICS
  # ==========================================================================
  /api/admin/dashboard:
    get:
      tags: [Admin – Analytics]
      summary: Key dashboard metrics
      description: |
        Returns MAU, new registrations, opportunity counts, email subscriptions,
        and month-over-month growth rate. **Admin only.**
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: object
                      opportunities:
                        type: object
                      email:
                        type: object
  /api/admin/analytics/usage:
    get:
      tags: [Admin – Analytics]
      summary: Daily/weekly usage trends
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          description: Number of days to look back (7–90)
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Usage trend data
  /api/admin/analytics/revenue:
    get:
      tags: [Admin – Analytics]
      summary: Subscription revenue breakdown
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revenue analytics
  /api/admin/analytics/features:
    get:
      tags: [Admin – Analytics]
      summary: Feature usage statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Feature usage
  /api/admin/analytics/search-trends:
    get:
      tags: [Admin – Analytics]
      summary: Popular searches and NAICS codes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Search trends

  # ==========================================================================
  # ADMIN – SUBSCRIPTIONS
  # ==========================================================================
  /api/admin/subscriptions:
    get:
      tags: [Admin – Subscriptions]
      summary: List all subscriptions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Subscription list with MRR summary
  /api/admin/subscriptions/{userId}:
    get:
      tags: [Admin – Subscriptions]
      summary: Get user subscription details
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription detail
  /api/admin/subscriptions/{userId}/extend-trial:
    post:
      tags: [Admin – Subscriptions]
      summary: Extend trial for a prospect
      description: Re-activates the user account. Use this during demo calls.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 30
                reason:
                  type: string
      responses:
        "200":
          description: Trial extended
  /api/admin/subscriptions/{userId}/issue-refund:
    post:
      tags: [Admin – Subscriptions]
      summary: Log a refund request
      description: |
        Records the refund request in the audit log.
        Process the actual refund in the **Stripe Dashboard**.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                reason:
                  type: string
      responses:
        "200":
          description: Refund request logged

  # ==========================================================================
  # ADMIN – SYSTEM
  # ==========================================================================
...
  /api/admin/logs:
    get:
      tags: [Admin – System]
      summary: System logs reference
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Log access instructions
  /api/admin/system/restart:
    post:
      tags: [Admin – System]
      summary: Initiate graceful server restart
      description: |
        Sends a `SIGTERM`-equivalent exit after flushing the response.
        Your process manager (PM2 / Render) should restart automatically.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Restart initiated

  # ==========================================================================
  # SYSTEM
  # ==========================================================================
...